Increment/Decrement to change and save. Read more
The increment and decrement methods do not fire the model Observer methods.

All GameTransactions update the 'profit' of the GameType (CG, Tournament).  When this 'profit' column is updated
the user's Bankroll needs to be updated as well, a double chaining.

The GameType Observer watching over its profit could not see that it was was updated as the GameTransactionObservers used to
change the profit was using increment/decrement which was not firing the GameTypeObserver model events.
Changed it to modify the profit attribute of the GameType and call the save() method directly.
The GameType Observer updated methods now fires, this includes the logic for updating the user's bankroll