Refactor Bankroll addition and withdrawls.  Read notes.
Adding and Withdrawing from a bankroll were separate functions where withdrawing just * -1 when just before creating a BankrollTransaction
Refactor to a single createBankrollTransaction() in User that accepts negative values for withdrawls instead.
This is so there was no need for two different axios endpoints in the frontend (/bankroll/add and /bankroll/withdraw)
and now everything can be done from one endpoint where we supply positive amount for addition, and negative amount for withdrawals, which is determined on the front end.
Withdrawals were stored as negative amount in the database anyway, so now we have less functions to maintain.