Fixed Chrome dev tool erros on delete. READ MORE
When deleting a Session, it tries to it tries to rerender the Session.vue/ TransactionSummary.vue DOM even though we redirect to Sessions.vue.
This results in trying to display all currency values using session.currency/transaction.currency which is undefined and so get lots of errors in Chrome Dev Tools.